\name{explore_paper}
\alias{explore_paper}
\title{explore_paper}
\usage{
explore_paper(url = 'http..... )
explore_paper(url = NULL, title,res )
}
\description{
returns tible with research papers on queried paper 
}
\examples{

papers <- explore_paper('Biochemistry')

explore_paper <- function(url = NULL, .title = 'Introduction', res = papers)

### Need to add more examples and stuff 
}
