\name{list.topics}
\alias{list.topics}
\title{list.topics}
\description{
Returns list of currently supported Nature subject areas
}

\usage{
list.topics()
}

\value{character vector of supported subject areas}

\examples{
# Supported subject areas
list.topics()
  [1] "Biochemistry"                                  
  [2] "Biological techniques"                         
  [3] "Biophysics"                                    
  [4] "Biotechnology" 
...
}
